Added caching to `get_soup` function + switched to `lxml`
Profiling showed that a significant amount of time was spent
within the `utils.get_soup` function.

Part of this was due to soup being produced for the same URL
multiple times; caching results within a dict resolved this.

Additional time was spent parsing HTML with the `html5lib` backend.
This backend is very lenient, but also quite slow. Since Real
Python has well formatted HTML, we do not need a lenient parser;
switched to `lxml` instead.